+++ frontend/client/letsencrypt.php
@@ -30,23 +30,34 @@ use PDO;
  * Functions
  */
 
+function letsencrypt_statusIcon($status) {
+    $statusIcon = 'error';
+    if ($status == 'ok') {
+        $statusIcon = 'ok';
+    } elseif ($status == 'disabled') {
+        $statusIcon = 'disabled';
+    } elseif (in_array(
+        $status,
+        array('toadd', 'tochange', 'todelete', 'torestore', 'toenable', 'todisable'))
+    ) {
+        $statusIcon = 'reload';
+    }
+    return $statusIcon;
+}
+
 /**
- * Generate page
+ * Generate domains
  *
  * @param $tpl TemplateEngine
  * @return void
  */
-function letsencrypt_generatePage($tpl)
+function letsencrypt_generateDomains($tpl)
 {
-
-    //SELECT * FROM `letsencrypt` LEFT JOIN `domain` ON letsencrypt.domain_id = domain.domain_id
-    // SELECT letsencrypt_id, cert_name, http_forward, status, state
-    // FROM letsencrypt
-    // WHERE admin_id = ?
     $stmt = exec_query(
         '
             SELECT domain.domain_id AS domain_id, domain_name, domain_admin_id, http_forward, status, state
-            FROM domain LEFT JOIN letsencrypt ON (
+            FROM domain
+            LEFT JOIN letsencrypt ON (
                 domain.domain_id=letsencrypt.domain_id
             )
             WHERE domain_admin_id = ?
@@ -54,46 +65,136 @@ function letsencrypt_generatePage($tpl)
         array($_SESSION['user_id'])
     );
 
-    if ($stmt->rowCount()) {
-        while ($row = $stmt->fetchRow(PDO::FETCH_ASSOC)) {
-            if (!$row['status']) {
-                $row['status'] = 'disabled';
-            }
-            if ($row['status'] == 'ok') {
-                $statusIcon = 'ok';
-            } elseif ($row['status'] == 'disabled') {
-                $statusIcon = 'disabled';
-            } elseif (in_array(
-                $row['status'],
-                array('toadd', 'tochange', 'todelete', 'torestore', 'toenable', 'todisable'))
-            ) {
-                $statusIcon = 'reload';
-            } else {
-                $statusIcon = 'error';
-            }
-
-            $type = 'domain';
-            $id = $row['domain_id'];
-            $domain_name = decode_idna($row['domain_name']);
-
-            $tpl->assign(array(
-                'DOMAIN_NAME'      => decode_idna($row['domain_name']),
-                'ID'               => $row['domain_id'],
-                'NOTE'             => $row['state'] ? $row['state'] : '',
-                'EDIT'             => tr('Edit'),
-                'EDIT_LINK'        => 'letsencrypt_edit.php?type=' . $type . '&id=' . $id,
-                'STATUS'           => translate_dmn_status($row['status']), // TODO Improve the translation for ssl CP 2017-07
-                'STATUS_ICON'      => $statusIcon,
-                'HTTP_FORWARD'     => $row['http_forward'] ? tr('yes') : tr('no'),
-                'HTTP_FORWRD_ICON' => $row['http_forward'] ? 'check' : '', // TODO
-            ));
-            $tpl->parse('DOMAINKEY_ITEM', 'domainkey_item'); // TODO
+    while ($row = $stmt->fetchRow(PDO::FETCH_ASSOC)) {
+        if (!$row['status']) {
+            $row['status'] = 'disabled';
         }
-    } else {
-        $tpl->assign('CUSTOMER_LIST', '');
-        set_page_message(tr('No domain with LetsEncrypt support has been found.'), 'static_info');
+        $statusIcon = letsencrypt_statusIcon($row['status']);
+        $type = 'domain';
+        $id = $row['domain_id'];
+
+        $tpl->assign(array(
+            'DOMAIN_NAME'      => decode_idna($row['domain_name']),
+            'ID'               => $row['domain_id'],
+            'NOTE'             => $row['state'] ? $row['state'] : '',
+            'EDIT'             => tr('Edit'),
+            'EDIT_LINK'        => 'letsencrypt_edit.php?type=' . $type . '&id=' . $id,
+            'STATUS'           => translate_dmn_status($row['status']), // TODO Improve the translation for ssl CP 2017-07
+            'STATUS_ICON'      => $statusIcon,
+            'HTTP_FORWARD'     => $row['http_forward'] ? tr('yes') : tr('no'),
+            'HTTP_FORWRD_ICON' => $row['http_forward'] ? 'check' : '', // TODO
+        ));
+        $tpl->parse('DOMAIN_ITEM', '.domain_item'); // TODO
     }
 }
+
+/**
+ * Generate aliases
+ *
+ * @param $tpl TemplateEngine
+ * @return void
+ */
+ function letsencrypt_generateAliases($tpl)
+ {
+     $stmt = exec_query(
+         '
+             SELECT domain_aliasses.alias_id AS alias_id, alias_name, domain_admin_id, http_forward, status, state
+             FROM domain
+             INNER JOIN domain_aliasses ON (domain.domain_id = domain_aliasses.domain_id)
+             LEFT JOIN letsencrypt ON (
+                 domain_aliasses.alias_id=letsencrypt.alias_id
+             )
+             WHERE domain_admin_id = ?
+         ',
+         array($_SESSION['user_id'])
+     );
+ 
+    if (!$stmt->rowCount()) {
+        $tpl->assign(array(
+            'ALS_MSG' => tr('You do not have domain aliases.'),
+            'ALS_LIST' => ''
+        ));
+        return;
+    }
+
+    while ($row = $stmt->fetchRow(PDO::FETCH_ASSOC)) {
+        if (!$row['status']) {
+            $row['status'] = 'disabled';
+        }
+        $statusIcon = letsencrypt_statusIcon($row['status']);
+        $type = 'alias';
+        $id = $row['alias_id'];
+
+        $tpl->assign(array(
+            'DOMAIN_NAME'      => decode_idna($row['alias_name']),
+            'ID'               => $row['alias_id'],
+            'NOTE'             => $row['state'] ? $row['state'] : '',
+            'EDIT'             => tr('Edit'),
+            'EDIT_LINK'        => 'letsencrypt_edit.php?type=' . $type . '&id=' . $id,
+            'STATUS'           => translate_dmn_status($row['status']), // TODO Improve the translation for ssl CP 2017-07
+            'STATUS_ICON'      => $statusIcon,
+            'HTTP_FORWARD'     => $row['http_forward'] ? tr('yes') : tr('no'),
+            'HTTP_FORWRD_ICON' => $row['http_forward'] ? 'check' : '', // TODO
+        ));
+        $tpl->parse('ALS_ITEM', '.als_item'); // TODO
+    }
+    $tpl->assign('ALS_MESSAGE', '');
+ }
+ 
+/**
+ * Generate subdomains
+ *
+ * @param $tpl TemplateEngine
+ * @return void
+ */
+ function letsencrypt_generateSubdomains($tpl)
+ {
+     $stmt = exec_query(
+         '
+             SELECT subdomain.subdomain_id AS subdomain_id, subdomain_name, domain_name, domain_admin_id, http_forward, status, state
+             FROM domain
+             INNER JOIN subdomain ON (domain.domain_id = subdomain.domain_id)
+             LEFT JOIN letsencrypt ON (
+                 subdomain.subdomain_id=letsencrypt.subdomain_id
+             )
+             WHERE domain_admin_id = ?
+         ',
+         array($_SESSION['user_id'])
+     );
+ 
+    if (!$stmt->rowCount()) {
+        $tpl->assign(array(
+            'SUB_MSG' => tr('You do not have subdomains.'),
+            'SUB_LIST' => ''
+        ));
+        return;
+    }
+
+    while ($row = $stmt->fetchRow(PDO::FETCH_ASSOC)) {
+        if (!$row['status']) {
+            $row['status'] = 'disabled';
+        }
+        $statusIcon = letsencrypt_statusIcon($row['status']);
+        $type = 'subdomain';
+        $id = $row['subdomain_id'];
+
+        $tpl->assign(array(
+            'DOMAIN_NAME'      => decode_idna($row['subdomain_name'] . '.' . $row['domain_name']),
+            'ID'               => $row['subdomain_id'],
+            'NOTE'             => $row['state'] ? $row['state'] : '',
+            'EDIT'             => tr('Edit'),
+            'EDIT_LINK'        => 'letsencrypt_edit.php?type=' . $type . '&id=' . $id,
+            'STATUS'           => translate_dmn_status($row['status']), // TODO Improve the translation for ssl CP 2017-07
+            'STATUS_ICON'      => $statusIcon,
+            'HTTP_FORWARD'     => $row['http_forward'] ? tr('yes') : tr('no'),
+            'HTTP_FORWRD_ICON' => $row['http_forward'] ? 'check' : '', // TODO
+        ));
+        $tpl->parse('SUB_ITEM', '.sub_item'); // TODO
+    }
+    $tpl->assign('SUB_MESSAGE', '');
+ }
+ 
+
 
 /***********************************************************************************************************************
  * Main
@@ -108,23 +209,40 @@ if (!LetsEncrypt::customerHasLetsEncrypt(intval($_SESSION['user_id']))) {
 
 $tpl = new TemplateEngine();
 $tpl->define_dynamic(array(
-    'layout'         => 'shared/layouts/ui.tpl',
-    'page'           => '../../plugins/LetsEncrypt/themes/default/view/client/letsencrypt.tpl',
-    'page_message'   => 'layout',
-    'customer_list'  => 'page',
-    'domainkey_item' => 'customer_list'
+    'layout'                     => 'shared/layouts/ui.tpl',
+    'page'                       => '../../plugins/LetsEncrypt/themes/default/view/client/letsencrypt.tpl',
+    'page_message'               => 'layout',
+    'domain_list'                => 'page',
+    'domain_item'                => 'domain_list',
+    'domain_status_reload_true'  => 'domain_item',
+    'domain_status_reload_false' => 'domain_item',
+    'domain_aliases_block'       => 'page',
+    'als_message'                => 'domain_aliases_block',
+    'als_list'                   => 'domain_aliases_block',
+    'als_item'                   => 'als_list',
+    'als_status_reload_true'     => 'als_item',
+    'als_status_reload_false'    => 'als_item',
+    'subdomains_block'           => 'page',
+    'sub_message'                => 'subdomains_block',
+    'sub_list'                   => 'subdomains_block',
+    'sub_item'                   => 'sub_list',
 ));
 $tpl->assign(array(
-    'TR_PAGE_TITLE'   => tr('Customers / LetsEncrypt'),
-    'TR_ACTION'       => tr('Actions'),
-    'TR_DOMAIN_NAME'  => tr('Domain'),
-    'TR_HTTP_FORWARD' => tr('Forward to SSL'),
-    'TR_NOTE'         => tr('Notes'),
-    'TR_STATUS'       => tr('Status')
+    'TR_PAGE_TITLE'     => tr('Customers / LetsEncrypt'),
+    'TR_ACTION'         => tr('Actions'),
+    'TR_DOMAINS'        => tr('Domains'),
+    'TR_DOMAIN_ALIASES' => tr('Aliases'),
+    'TR_SUBDOMAINS'     => tr('Subdomains'),
+    'TR_DOMAIN_NAME'    => tr('Domain'),
+    'TR_HTTP_FORWARD'   => tr('Forward to SSL'),
+    'TR_NOTE'           => tr('Notes'),
+    'TR_STATUS'         => tr('Status')
 ));
 
 generateNavigation($tpl);
-letsencrypt_generatePage($tpl);
+letsencrypt_generateDomains($tpl);
+letsencrypt_generateAliases($tpl);
+letsencrypt_generateSubdomains($tpl);
 generatePageMessage($tpl);
 
 $tpl->parse('LAYOUT_CONTENT', 'page');